Avoid websocket heartbeat wrapper import-time cost
